JavaLog moved to org.eclipse.persistence.logging.jul package.
(cherry picked from commit 6f125053b38b6cc313ec48b83845257393a39ac8)
