+++ bundles/eclipselink/src/main/java/module-info.java
@@ -74,6 +74,7 @@ module eclipselink {
     exports org.eclipse.persistence.history;
     exports org.eclipse.persistence.indirection;
     exports org.eclipse.persistence.logging;
+    exports org.eclipse.persistence.logging.jul;
     exports org.eclipse.persistence.mappings;
     exports org.eclipse.persistence.mappings.converters;
     exports org.eclipse.persistence.mappings.foundation;
