+++ foundation/org.eclipse.persistence.core/src/main/java/module-info.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021, 2024 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2021, 2025 Oracle and/or its affiliates. All rights reserved.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0 which is available at
@@ -66,6 +66,7 @@ module org.eclipse.persistence.core {
     exports org.eclipse.persistence.history;
     exports org.eclipse.persistence.indirection;
     exports org.eclipse.persistence.logging;
+    exports org.eclipse.persistence.logging.jul;
     exports org.eclipse.persistence.mappings;
     exports org.eclipse.persistence.mappings.converters;
     exports org.eclipse.persistence.mappings.foundation;
